Cambios en el metodo delete de Linked List, cambiar o ajustar en cada DAO, SERVICES y API

+++ trabajo_final/src/main/java/com/example/controller/dao/services/UsuarioService.java
@@ -19,7 +19,8 @@ public class UsuarioService {
         return userDao.update();
     }
     public Boolean delete(Integer id) throws Exception {
-        return userDao.delete(id); 
+//        return userDao.delete(id);
+        return userDao.delete(id.intValue());
     }
